+++ source/spritetools.c
@@ -8,7 +8,7 @@
 
 /* Inits all modules */
 /* Returns 1 on success, 0 on failure */
-int ST_Init(void)
+u8 ST_Init(void)
 {
   if (!ST_DebugInit())
     return 0;
@@ -24,7 +24,7 @@ int ST_Init(void)
 
 /* Cleans up all modules */
 /* Returns 1 on success, 0 on failure */
-int ST_Fini(void)
+u8 ST_Fini(void)
 {
   if (!ST_DebugFini())
     return 0;
